test for swap records, convert labels from char* to ulong, remove rel flags, change slot size to 32 byte, get page size from os, and more

+++ src/io/disk_file.h
@@ -11,13 +11,14 @@
 #define DISK_FILE_H
 
 #include <stdbool.h>
+#include <stddef.h>
 #include <stdio.h>
 
 typedef struct
 {
     char*  file_name;
     FILE*  file;
-    long   file_size;
+    size_t file_size;
     size_t num_pages;
     size_t read_count;
     size_t write_count;
